read from all sensors at the same time

+++ sensors/BME280.h
@@ -8,7 +8,7 @@
 
 int BME280Init(bmx280_t* dev, i2c_t i2c_dev, uint8_t i2c_addr);
 float readTemperature(bmx280_t* dev);
-uint32_t readPressure(bmx280_t* dev);
-double readHumidity(bmx280_t* dev);
+float readPressure(bmx280_t* dev);
+float readHumidity(bmx280_t* dev);
 
 #endif
